fix: gamesContext created
