+++ src/components/Games/Games.js
@@ -1,29 +1,14 @@
-import React, { useEffect, useState } from "react";
+import React, { useEffect, useContext } from "react";
 import { Game } from "./Game/Game";
 import { Grid } from "@mui/material";
-import { firestore } from "../../firebaseConf";
-import { collection, getDocs } from "firebase/firestore";
+import { gamesContext } from "../../context/gamesContext";
 
 export function Games() {
-  const [games, setGames] = useState([]);
-
-  const getGames = async () => {
-    const docs = [];
-    const querySnapshot = await getDocs(collection(firestore, "games"));
-    console.log(querySnapshot);
-    querySnapshot.forEach((doc) => {
-      console.log(doc.data());
-      console.log(doc.id);
-      docs.push({ ...doc.data(), id: doc.id });
-    });
-
-    console.log(docs);
-    setGames(docs);
-  };
+  const { games, getGames } = useContext(gamesContext);
 
   useEffect(() => {
     getGames();
-  }, []);
+  }, [getGames]);
 
   return (
     <Grid
