+++ src/components/Home.js
@@ -1,6 +1,11 @@
 import React from "react";
 import { Games } from "./Games/Games.js";
+import { GamesProvider } from "../context/gamesContext.js";
 
 export function Home() {
-  return <Games />;
+  return (
+    <GamesProvider>
+      <Games />
+    </GamesProvider>
+  );
 }
